+++ src/reactpy_router/routers.py
@@ -4,16 +4,16 @@ from __future__ import annotations
 
 from dataclasses import replace
 from logging import getLogger
-from typing import Any, Iterator, Literal, Sequence
+from typing import Any, Iterator, Literal, Sequence, cast
 
 from reactpy import component, use_memo, use_state
 from reactpy.backend.hooks import ConnectionContext, use_connection
 from reactpy.backend.types import Connection, Location
-from reactpy.core.types import VdomDict
-from reactpy.types import ComponentType
+from reactpy.core.component import Component
+from reactpy.types import ComponentType, VdomDict
 
-from reactpy_router.components import History
-from reactpy_router.hooks import _route_state_context, _RouteState
+from reactpy_router.components import FirstLoad, History
+from reactpy_router.hooks import RouteState, _route_state_context
 from reactpy_router.resolvers import StarletteResolver
 from reactpy_router.types import CompiledRoute, Resolver, Router, RouteType
 
@@ -24,15 +24,27 @@ _logger = getLogger(__name__)
 def create_router(resolver: Resolver[RouteType]) -> Router[RouteType]:
     """A decorator that turns a resolver into a router"""
 
-    def wrapper(*routes: RouteType) -> ComponentType:
+    def wrapper(*routes: RouteType) -> Component:
         return router(*routes, resolver=resolver)
 
     return wrapper
 
 
-browser_router = create_router(StarletteResolver)
-"""This is the recommended router for all ReactPy Router web projects.
-It uses the JavaScript DOM History API to manage the history stack."""
+_starlette_router = create_router(StarletteResolver)
+
+
+def browser_router(*routes: RouteType) -> Component:
+    """This is the recommended router for all ReactPy-Router web projects.
+    It uses the JavaScript [History API](https://developer.mozilla.org/en-US/docs/Web/API/History_API)
+    to manage the history stack.
+
+    Args:
+        *routes (RouteType): A list of routes to be rendered by the router.
+
+    Returns:
+        A router component that renders the given routes.
+    """
+    return _starlette_router(*routes)
 
 
 @component
@@ -47,6 +59,7 @@ def router(
 
     old_conn = use_connection()
     location, set_location = use_state(old_conn.location)
+    first_load, set_first_load = use_state(True)
 
     resolvers = use_memo(
         lambda: tuple(map(resolver, _iter_routes(routes))),
@@ -59,7 +72,7 @@ def router(
         route_elements = [
             _route_state_context(
                 element,
-                value=_RouteState(set_location, params),
+                value=RouteState(set_location, params),
             )
             for element, params in match
         ]
@@ -70,8 +83,15 @@ def router(
             if location != new_location:
                 set_location(new_location)
 
+        def on_first_load(event: dict[str, Any]) -> None:
+            """Callback function used within the JavaScript `FirstLoad` component."""
+            if first_load:
+                set_first_load(False)
+                on_history_change(event)
+
         return ConnectionContext(
-            History({"onHistoryChange": on_history_change}),  # type: ignore[return-value]
+            History({"onHistoryChangeCallback": on_history_change}),  # type: ignore[return-value]
+            FirstLoad({"onFirstLoadCallback": on_first_load}) if first_load else "",
             *route_elements,
             value=Connection(old_conn.scope, location, old_conn.carrier),
         )
@@ -86,6 +106,18 @@ def _iter_routes(routes: Sequence[RouteType]) -> Iterator[RouteType]:
         yield parent
 
 
+def _add_route_key(match: tuple[Any, dict[str, Any]], key: str | int) -> Any:
+    """Add a key to the VDOM or component on the current route, if it doesn't already have one."""
+    element, _params = match
+    if hasattr(element, "render") and not element.key:
+        element = cast(ComponentType, element)
+        element.key = key
+    elif isinstance(element, dict) and not element.get("key", None):
+        element = cast(VdomDict, element)
+        element["key"] = key
+    return match
+
+
 def _match_route(
     compiled_routes: Sequence[CompiledRoute],
     location: Location,
@@ -97,12 +129,14 @@ def _match_route(
         match = resolver.resolve(location.pathname)
         if match is not None:
             if select == "first":
-                return [match]
+                return [_add_route_key(match, resolver.key)]
 
             # Matching multiple routes is disabled since `react-router` no longer supports multiple
             # matches via the `Route` component. However, it's kept here to support future changes
             # or third-party routers.
-            matches.append(match)  # pragma: no cover
+            # TODO: The `resolver.key` value has edge cases where it is not unique enough to use as
+            # a key here. We can potentially fix this by throwing errors for duplicate identical routes.
+            matches.append(_add_route_key(match, resolver.key))  # pragma: no cover
 
     if not matches:
         _logger.debug("No matching route found for %s", location.pathname)