+++ tests/conftest.py
@@ -1,6 +1,4 @@
-import asyncio
 import os
-import sys
 
 import pytest
 from playwright.async_api import async_playwright
@@ -18,27 +16,25 @@ def pytest_addoption(parser) -> None:
     )
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 async def display(backend, browser):
     async with DisplayFixture(backend, browser) as display_fixture:
         display_fixture.page.set_default_timeout(10000)
         yield display_fixture
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 async def backend():
     async with BackendFixture() as backend_fixture:
         yield backend_fixture
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 async def browser(pytestconfig):
     async with async_playwright() as pw:
         yield await pw.chromium.launch(headless=True if GITHUB_ACTIONS else pytestconfig.getoption("headless"))
 
 
-@pytest.fixture
-def event_loop_policy(request):
-    if sys.platform == "win32":
-        return asyncio.WindowsProactorEventLoopPolicy()
-    return asyncio.get_event_loop_policy()
+@pytest.fixture(scope="session")
+def anyio_backend():
+    return "asyncio"